[close-failing-tests-stale-issues-task] Moved to notify
